Split UI Class into Different File
